Change how the location module is initialized
