+++ connect-location/src/main/java/com/ifttt/location/AwarenessEnterReceiver.java
@@ -4,7 +4,7 @@ import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 
-final class AwarenessEnterReceiver extends BroadcastReceiver {
+public final class AwarenessEnterReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(Context context, Intent intent) {
